move the vendored functions to blackcompat
